nexus/atlas_nexus_robot/backend/requirements.txt
# ATLAS NEXUS Robot Backend - Dependencies

# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# Computer Vision & Camera
opencv-python==4.8.1.78
opencv-contrib-python==4.8.1.78
pillow>=9.0.0
numpy==1.24.3

# AI & Machine Learning
tensorflow==2.15.0
torch==2.1.1
torchvision==0.16.1
scikit-learn==1.3.2
transformers==4.36.0

# Robotics & Hardware
serial==0.0.97
pyserial==3.5

# Communication & Networking
websockets==12.0
aiohttp==3.9.1
requests==2.31.0
paho-mqtt==1.6.1

# Configuration & Data
pyyaml==6.0.1
configparser==6.0.0
python-dotenv==1.0.0
sqlalchemy==2.0.23

# Utilities & Tools
schedule==1.2.0
watchdog==3.0.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Development
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# AI and Reasoning
langchain>=0.1.0
langchain-community>=0.0.10
langchain-openai>=0.0.5
chromadb>=0.4.22
sentence-transformers>=2.2.0
openai>=1.12.0
tiktoken>=0.5.2

# Additional Computer Vision
scipy==1.11.4
imageio==2.31.6
scikit-image==0.22.0

# Identity Recognition System
# face-recognition==1.3.0  # Requires dlib - problematic on Windows
deepface==0.0.79
# resemblyzer==0.1.1.dev0  # Development version - may be unstable
librosa==0.10.1

# Database for User Profiles
sqlalchemy==2.0.23
